code/imageConverter/requirements.txt
cycler==0.11.0
fonttools==4.29.1
imageio==2.16.1
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.7
numpy==1.22.2
opencv-python==4.5.5.62
packaging==21.3
Pillow==9.0.1
polarTransform==2.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.2.0
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
tifffile==2022.2.9
